changed database course grade attribute into float

+++ db/migrate/20140517055326_add_course_grade_to_course.rb
@@ -1,5 +1,5 @@
 class AddCourseGradeToCourse < ActiveRecord::Migration
   def change
-  	add_column :courses, :grade, :integer, :default => 0
+  	add_column :courses, :grade, :float, :default => 0.0
   end
 end
